gles-user-module: fix occasional qtbase configure build issue
Building the meta-qt5 component qtbase occasionally failures in the
do_configure() task with the error:
make: *** [opengles2] Error 1
OpenGL ES 2.0 disabled.
The OpenGL ES 2.0 functionality test failed!

This is a build order problem caused by an incomplete build dependency
chain. Linking opengles2 against libIMGegl.so (the virtual/libgles2
provider) has a dependency on libgbm.so and libwayland-kms.so so they
must be built first. qtbase does not have many gfx components in its
build depedency list and therefore may build earlier, which in turn
results in the error.

Simply adding libgbm and wayland-kms to the gles-user-module DEPENDS
results in a circular build depedency. This is a workaround to the
problem whilst the build depdency issue is resolved. For it to work
add packagegroup-graphics-libegl to your IMAGE_INSTALL.

This fixes the qtbase do_configure() build error reported in GDP-590 [1].

[1] https://at.projects.genivi.org/jira/browse/GDP-590