conf: select gcc version 5.3 in sample conf
The Linaro toolchain has a known build failure when rpm is
used with gcc 5.2. The gcc v5.2 libgcc recipe will fail in
the do_package_write_rpm() task. This is issue #2431 in the
Linaro bugzilla [1]. It is caused by the '-2' that is appended
to the end of the release string. To avoid the issue update
the local.conf to use gcc 5.3 so rpm can be used.

[1] https://bugs.linaro.org/show_bug.cgi?id=2431

Signed-off-by: Stephen Lawrence <stephen.lawrence@renesas.com>